requirements.txt
boto3
simplejson>=2.1
python-dateutil>=2.1
unittest2>=0.5.1
cryptography
urllib3
